# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/get_secondary_service/{sortOrder}": {
            "get": {
                "tags": [
                    "Secondary Services"
                ],
                "summary": "Secondary service list",
                "operationId": "08f402262bddee3a269e16fbc4e93a31",
                "parameters": [
                    {
                        "name": "sortOrder",
                        "in": "query",
                        "description": "Primary service id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Primary Service List"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                }
            }
        },
        "/api/auth/store_secondary_service": {
            "post": {
                "tags": [
                    "Secondary Services"
                ],
                "summary": "Store secondary service",
                "operationId": "443ed57ebe9636a24daa9f7aff80d785",
                "parameters": [
                    {
                        "name": "primary_service_id",
                        "in": "query",
                        "description": "Primary service Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "secondaryName",
                        "in": "query",
                        "description": "Secondary service name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "business_key",
                        "in": "query",
                        "description": "Secondary service business_key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "urlIcon": {
                                        "description": "Image file",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Seondary Service Created"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/auth/update_secondary_service/{id}": {
            "post": {
                "tags": [
                    "Secondary Services"
                ],
                "summary": "Update secondary service",
                "operationId": "e2b857a7c92236d20683667b55002263",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Secondary service Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "primary_service_id",
                        "in": "query",
                        "description": "Primary service Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "secondaryName",
                        "in": "query",
                        "description": "Secondary service name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "business_key",
                        "in": "query",
                        "description": "Secondary service business_key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "urlIcon": {
                                        "description": "Image file",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Seondary Service Updated"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/auth/delete_secondary_service/{id}": {
            "get": {
                "tags": [
                    "Secondary Services"
                ],
                "summary": "Delete secondary service",
                "operationId": "7e4f5bafe7ae26bab77acce3342225ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Primary service id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Secondary Service"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}